update evaluation for unify coco and voc
