+++ configs/ssd/ssd_mobilenet_v1_voc.yml
@@ -95,7 +95,7 @@ TrainReader:
 EvalReader:
   inputs_def:
     image_shape: [3, 300, 300]
-    fields: ['image', 'gt_bbox', 'gt_class', 'im_shape', 'im_id', 'is_difficult']
+    fields: ['image' 'im_shape', 'im_id']
   dataset:
     !VOCDataSet
     anno_path: test.txt
