+++ configs/ssd/ssd_vgg16_300_voc.yml
@@ -98,7 +98,7 @@ TrainReader:
 EvalReader:
   inputs_def:
     image_shape: [3, 300, 300]
-    fields: ['image', 'gt_bbox', 'gt_class', 'im_shape', 'im_id', 'is_difficult']
+    fields: ['image' 'im_shape', 'im_id']
   dataset:
     !VOCDataSet
     anno_path: test.txt
