+++ configs/ssd/ssd_vgg16_512_voc.yml
@@ -102,7 +102,7 @@ TrainReader:
 EvalReader:
   inputs_def:
     image_shape: [3, 512, 512]
-    fields: ['image', 'gt_bbox', 'gt_class', 'im_shape', 'im_id', 'is_difficult']
+    fields: ['image' 'im_shape', 'im_id']
   dataset:
     !VOCDataSet
     anno_path: test.txt
