+++ configs/yolov3_mobilenet_v1_voc.yml
@@ -70,9 +70,6 @@ TrainReader:
     with_background: false
 
 EvalReader:
-  inputs_def:
-    fields: ['image', 'im_size', 'im_id', 'gt_bbox', 'gt_class', 'is_difficult']
-    num_max_boxes: 50
   dataset:
     !VOCDataSet
     dataset_dir: dataset/voc
