+++ ppdet/data/source/oid.py
@@ -21,12 +21,10 @@ import os
 import sys
 import numpy as np
 
-from .coco_eval import bbox2out
-
 import logging
 logger = logging.getLogger(__name__)
 
-__all__ = ['bbox2out', 'get_category_info']
+__all__ = ['get_category_info']
 
 
 def get_category_info(anno_file=None,
