+++ slim/extensions/distill_pruned_model/distill_pruned_model.py
@@ -33,8 +33,8 @@ import paddle
 from paddle import fluid
 from ppdet.core.workspace import load_config, merge_config, create
 from ppdet.data.reader import create_reader
-from ppdet.utils.eval_utils import parse_fetches, eval_results, eval_run
-from ppdet.utils.stats import TrainingStats
+from ppdet.evaluation.evaluate import Evaluation
+from ppdet.utils.stats import TrainingStats, parse_fetches
 from ppdet.utils.cli import ArgsParser
 from ppdet.utils.check import check_gpu, check_version, check_config, enable_static_mode
 import ppdet.utils.checkpoint as checkpoint
@@ -304,11 +304,7 @@ def main():
     compiled_eval_prog = fluid.CompiledProgram(eval_prog)
 
     # parse eval fetches
-    extra_keys = []
-    if cfg.metric == 'COCO':
-        extra_keys = ['im_info', 'im_id', 'im_shape']
-    if cfg.metric == 'VOC':
-        extra_keys = ['gt_bbox', 'gt_class', 'is_difficult']
+    extra_keys = ['im_info', 'im_id', 'im_shape']
     eval_keys, eval_values, eval_cls = parse_fetches(fetches, eval_prog,
                                                      extra_keys)
 
@@ -321,7 +317,7 @@ def main():
     best_box_ap_list = [0.0, 0]  #[map, iter]
     cfg_name = os.path.basename(FLAGS.config).split('.')[0]
     save_dir = os.path.join(cfg.save_dir, cfg_name)
-
+    eval = Evaluation(cfg, is_bbox_normalized=is_bbox_normalized)
     train_loader.start()
     for step_id in range(start_iter, cfg.max_iters):
         teacher_loss_np, distill_loss_np, loss_np, lr_np = exe.run(
@@ -341,16 +337,13 @@ def main():
             checkpoint.save(exe, distill_prog,
                             os.path.join(save_dir, save_name))
             # eval
-            results = eval_run(exe, compiled_eval_prog, eval_loader, eval_keys,
-                               eval_values, eval_cls, cfg)
-            resolution = None
-            box_ap_stats = eval_results(results, cfg.metric, cfg.num_classes,
-                                        resolution, is_bbox_normalized,
-                                        FLAGS.output_eval, map_type,
-                                        cfg['EvalReader']['dataset'])
+            results = eval.eval_run(exe, compiled_eval_prog, eval_loader,
+                                    eval_keys, eval_values)
 
-            if box_ap_stats[0] > best_box_ap_list[0]:
-                best_box_ap_list[0] = box_ap_stats[0]
+            box_ap_stats = eval.eval_results()
+
+            if box_ap_stats > best_box_ap_list[0]:
+                best_box_ap_list[0] = box_ap_stats
                 best_box_ap_list[1] = step_id
                 checkpoint.save(exe, distill_prog,
                                 os.path.join("./", "best_model"))
