+++ slim/prune/eval.py
@@ -28,9 +28,10 @@ import paddle.fluid as fluid
 from paddleslim.prune import Pruner
 from paddleslim.analysis import flops
 
-from ppdet.utils.eval_utils import parse_fetches, eval_run, eval_results, json_eval_results
+from ppdet.utils.stats import parse_fetches
 import ppdet.utils.checkpoint as checkpoint
 from ppdet.utils.check import check_gpu, check_version, check_config, enable_static_mode
+from ppdet.evaluation.evaluate import Evaluation
 
 from ppdet.data.reader import create_reader
 
@@ -83,6 +84,14 @@ def main():
     loader.set_sample_list_generator(reader)
 
     dataset = cfg['EvalReader']['dataset']
+    # whether output bbox is normalized in model output layer
+    is_bbox_normalized = False
+    if hasattr(model, 'is_bbox_normalized') and \
+            callable(model.is_bbox_normalized):
+        is_bbox_normalized = model.is_bbox_normalized()
+
+    eval = Evaluation(
+        cfg, is_bbox_normalized=is_bbox_normalized, eval_dir=FLAGS.output_eval)
 
     # eval already exists json file
     if FLAGS.json_eval:
@@ -90,8 +99,7 @@ def main():
             "In json_eval mode, PaddleDetection will evaluate json files in "
             "output_eval directly. And proposal.json, bbox.json and mask.json "
             "will be detected by default.")
-        json_eval_results(
-            cfg.metric, json_directory=FLAGS.output_eval, dataset=dataset)
+        eval.json_eval_results()
         return
 
     pruned_params = FLAGS.pruned_params
@@ -131,20 +139,9 @@ def main():
                          "please use tools/face_eval.py".format(cfg.metric))
     assert cfg.metric in ['COCO', 'VOC'], \
             "unknown metric type {}".format(cfg.metric)
-    extra_keys = []
 
-    if cfg.metric == 'COCO':
-        extra_keys = ['im_info', 'im_id', 'im_shape']
-    if cfg.metric == 'VOC':
-        extra_keys = ['gt_bbox', 'gt_class', 'is_difficult']
-
+    extra_keys = ['im_info', 'im_id', 'im_shape']
     keys, values, cls = parse_fetches(fetches, eval_prog, extra_keys)
-
-    # whether output bbox is normalized in model output layer
-    is_bbox_normalized = False
-    if hasattr(model, 'is_bbox_normalized') and \
-            callable(model.is_bbox_normalized):
-        is_bbox_normalized = model.is_bbox_normalized()
 
     sub_eval_prog = None
     sub_keys = None
@@ -173,30 +170,14 @@ def main():
     if 'Mask' in cfg.architecture:
         resolution = model.mask_head.resolution
 
-    results = eval_run(
-        exe,
-        compile_program,
-        loader,
-        keys,
-        values,
-        cls,
-        cfg,
-        sub_eval_prog,
-        sub_keys,
-        sub_values,
-        resolution=resolution)
+    results = eval.eval_run(exe, compile_program, loader, keys, values,
+                            sub_eval_prog, sub_keys, sub_values, resolution)
 
-    # if map_type not set, use default 11point, only use in VOC eval
-    map_type = cfg.map_type if 'map_type' in cfg else '11point'
-    eval_results(
-        results,
-        cfg.metric,
-        cfg.num_classes,
-        resolution,
-        is_bbox_normalized,
-        FLAGS.output_eval,
-        map_type,
-        dataset=dataset)
+    # evaluation
+    save_only = getattr(cfg, 'save_prediction_only', False)
+    if save_only:
+        eval.save_eval_json(FLAGS.output_eval)
+    eval.eval_results(classwise=FLAGS.classwise)
 
 
 if __name__ == '__main__':
@@ -226,6 +207,10 @@ if __name__ == '__main__':
         type=str,
         help="The ratios pruned iteratively for each parameter when calculating sensitivities."
     )
-
+    parser.add_argument(
+        "--classwise",
+        action='store_true',
+        default=False,
+        help="Whether to compute per-category AP")
     FLAGS = parser.parse_args()
     main()
