+++ slim/quantization/eval.py
@@ -26,10 +26,10 @@ if parent_path not in sys.path:
 import paddle
 import paddle.fluid as fluid
 
-from ppdet.utils.eval_utils import parse_fetches, eval_run, eval_results, json_eval_results
+from ppdet.utils.stats import parse_fetches
 import ppdet.utils.checkpoint as checkpoint
 from ppdet.utils.check import check_gpu, check_version, check_config, enable_static_mode
-
+from ppdet.evaluation.evaluate import Evaluation
 from ppdet.data.reader import create_reader
 
 from ppdet.core.workspace import load_config, merge_config, create
@@ -77,14 +77,21 @@ def main():
     # When iterable mode, set set_sample_list_generator(reader, place)
     loader.set_sample_list_generator(reader)
 
+    # whether output bbox is normalized in model output layer
+    is_bbox_normalized = False
+    if hasattr(model, 'is_bbox_normalized') and \
+            callable(model.is_bbox_normalized):
+        is_bbox_normalized = model.is_bbox_normalized()
+
+    eval = Evaluation(
+        cfg, is_bbox_normalized=is_bbox_normalized, eval_dir=FLAGS.output_eval)
     # eval already exists json file
     if FLAGS.json_eval:
         logger.info(
             "In json_eval mode, PaddleDetection will evaluate json files in "
             "output_eval directly. And proposal.json, bbox.json and mask.json "
             "will be detected by default.")
-        json_eval_results(
-            cfg.metric, json_directory=FLAGS.output_eval, dataset=dataset)
+        eval.json_eval_results()
         return
 
     assert cfg.metric != 'OID', "eval process of OID dataset \
@@ -95,20 +102,8 @@ def main():
                          "please use tools/face_eval.py".format(cfg.metric))
     assert cfg.metric in ['COCO', 'VOC'], \
             "unknown metric type {}".format(cfg.metric)
-    extra_keys = []
-
-    if cfg.metric == 'COCO':
-        extra_keys = ['im_info', 'im_id', 'im_shape']
-    if cfg.metric == 'VOC':
-        extra_keys = ['gt_bbox', 'gt_class', 'is_difficult']
-
+    extra_keys = ['im_info', 'im_id', 'im_shape']
     keys, values, cls = parse_fetches(test_fetches, eval_prog, extra_keys)
-
-    # whether output bbox is normalized in model output layer
-    is_bbox_normalized = False
-    if hasattr(model, 'is_bbox_normalized') and \
-            callable(model.is_bbox_normalized):
-        is_bbox_normalized = model.is_bbox_normalized()
 
     dataset = cfg['EvalReader']['dataset']
 
@@ -136,24 +131,15 @@ def main():
 
     compile_program = fluid.CompiledProgram(eval_prog).with_data_parallel()
 
-    results = eval_run(exe, compile_program, loader, keys, values, cls, cfg,
-                       sub_eval_prog, sub_keys, sub_values)
+    resolution = None
+    if 'Mask' in cfg.architecture or cfg.architecture == 'HybridTaskCascade':
+        resolution = model.mask_head.resolution
 
     # evaluation
-    resolution = None
-    if 'mask' in results[0]:
-        resolution = model.mask_head.resolution
-    # if map_type not set, use default 11point, only use in VOC eval
-    map_type = cfg.map_type if 'map_type' in cfg else '11point'
-    eval_results(
-        results,
-        cfg.metric,
-        cfg.num_classes,
-        resolution,
-        is_bbox_normalized,
-        FLAGS.output_eval,
-        map_type,
-        dataset=dataset)
+    results = eval.eval_run(exe, compile_program, loader, keys, values,
+                            sub_eval_prog, sub_keys, sub_values, resolution)
+
+    eval.eval_results(classwise=FLAGS.classwise)
 
 
 if __name__ == '__main__':
@@ -176,6 +162,10 @@ if __name__ == '__main__':
         type=str,
         help="Layers which name_scope contains string in not_quant_pattern will not be quantized"
     )
-
+    parser.add_argument(
+        "--classwise",
+        action='store_true',
+        default=False,
+        help="Whether to compute per-category AP")
     FLAGS = parser.parse_args()
     main()
